refactor: Move utilities from modules/ to utils/
