+++ gradience/ui/builtin_preset_row.py
@@ -19,7 +19,7 @@
 from gi.repository import Gtk, Adw
 
 from gradience.constants import rootdir
-from gradience.modules.utils import to_slug_case, buglog
+from gradience.utils.utils import to_slug_case, buglog
 
 
 @Gtk.Template(resource_path=f"{rootdir}/ui/builtin_preset_row.ui")
