+++ gradience/utils/README.md
@@ -1,3 +1,3 @@
-# Modules
+# Utilities
 
 Lots of modules are going to move to the Gradience-lib. We will only keep utilities here.
